+++ workflow/util/util.go
@@ -907,36 +907,3 @@ func GetNodeType(tmpl *wfv1.Template) wfv1.NodeType {
 	}
 	return ""
 }
-
-// MergeTo will do strategic merge the workflows
-// patch workflow will be merged into target workflow.
-// Overwrite flag indicates whether merge needs to overwrite the target value from patch
-func MergeTo(patch, target *wfv1.Workflow, overwrite bool) error {
-	if target == nil || patch == nil {
-		return nil
-	}
-
-	patchWfBytes, err := json.Marshal(patch)
-	if err != nil {
-		return err
-	}
-
-	targetWfByte, err := json.Marshal(target)
-	if err != nil {
-		return err
-	}
-	var mergedWfByte []byte
-	if overwrite {
-		mergedWfByte, err = strategicpatch.StrategicMergePatch(targetWfByte, patchWfBytes, wfv1.Workflow{})
-	} else {
-		mergedWfByte, err = strategicpatch.StrategicMergePatch(patchWfBytes, targetWfByte, wfv1.Workflow{})
-	}
-	if err != nil {
-		return err
-	}
-	err = json.Unmarshal(mergedWfByte, target)
-	if err != nil {
-		return err
-	}
-	return nil
-}
